Title:		Update main tester, little fix on ft_strcmp, libasm.h, Makefile Content:	Makefile: Remove flag -no-pie, use clang intead of gcc. 			ft_strcmp.s fix bug on same argument.

+++ libasm.h
@@ -6,7 +6,7 @@
 /*   By: cdai <cdai@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/11 15:10:26 by cdai              #+#    #+#             */
-/*   Updated: 2020/10/07 16:09:25 by cdai             ###   ########.fr       */
+/*   Updated: 2020/10/09 11:35:19 by cdai             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,11 @@
 #include <fcntl.h>
 #include <errno.h>
 
-int		ft_strlen(const char *s);
+size_t		ft_strlen(const char *s);
 char    *ft_strcpy(char *dest, const char *src);
 int     ft_strcmp(const char *s1, const char *s2);
-int     ft_write(int fd, char* buf, size_t	count);
-int     ft_read(int fd, char* buf, size_t count);
+ssize_t ft_write(int fd, char* buf, size_t	count);
+ssize_t ft_read(int fd, char* buf, size_t count);
 char    *ft_strdup(const char *str);
 
 #endif
